Fix qdbusperformance and qprocess benchmarks
The tests failed on 5.15 for different reasons than they do on dev, so
this cherry-pick is significantly modified. Only changes to qmake
configuration (absent on dev) were needed, but this reworking includes
such of the C++ code changes as are still a benefit, too.

The 5.15 failures were due to mis-configuring qmake: the qprocess
test's server was told it was a benchmark; the qdbusperformance
benchmark didn't know that was what it was.

Both tests called a QProcess::start() overload that was deprecated at
5.15 (even though the code still compiles just fine after removal of
that overload at 6.0), so silence the warning by passing an empty
QStringList().  Add the ".exe" suffix for QProcess's server path and
apply QFINDTESTDATA() to it.

Task-number: QTBUG-91713
Change-Id: I5ace23a5815575bbf88ea88e0b16afc7b8ba6a08
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit bf99c2b62bc3bc2dc10ba8f98bea0496b2026801)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>