QSet::erase - extend docs
Explicitly specify that calling this method for an empty set or with
an invalid iterator results in undefined behavior.

On a debug build an assert is triggered in such case, but on a release
build it will access the incorect index of an array.

Task-number: QTBUG-91736
Change-Id: Ibc3e91512a0ad9d9779a41083fedb8a91780380b
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit e5ec09ba868685ab2f943b39732a3cac018fedad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>